Takes the output of Stringtie and re-annotates the gene_id with ensembl ids
Stringtie assigns an MSTRG gene_id if there is a novel transcript isoform that doesn't
match the ref. This script ignores that, and if any of the transcripts match a ref ens id
it assigns that id to the gene.

Still has some faults - ie the GTF file needs to be sorted prior to running the script -
but it works...